Refactor: mapping method dto로 이동, 생성자로 처리 -> mapper 삭제

+++ src/main/java/weteam/backend/hashtag/HashtagService.java
@@ -9,10 +9,11 @@ import weteam.backend.config.handler.exception.BadRequestException;
 import weteam.backend.config.message.ExceptionMessage;
 import weteam.backend.hashtag.domain.Hashtag;
 import weteam.backend.hashtag.domain.MemberHashtag;
+import weteam.backend.hashtag.dto.AddHashtagDto;
 import weteam.backend.hashtag.dto.HashtagDto;
 import weteam.backend.hashtag.repository.HashtagRepository;
-import weteam.backend.hashtag.repository.MemberHashtagRepositorySupport;
 import weteam.backend.hashtag.repository.MemberHashtagRepository;
+import weteam.backend.hashtag.repository.MemberHashtagRepositorySupport;
 import weteam.backend.member.MemberService;
 import weteam.backend.member.domain.Member;
 
@@ -28,34 +29,33 @@ public class HashtagService {
     private final MemberHashtagRepositorySupport memberHashtagRepositorySupport;
     private final MemberService memberService;
 
-    public HashtagDto.Res save(final HashtagDto hashtagDto, final Long memberId) {
+    public HashtagDto create(final AddHashtagDto hashtagDto, final Long memberId) {
         Optional<Hashtag> data = hashTagRepository.findByName(hashtagDto.getName());
         Member member = memberService.findById(memberId);
 
         if (data.isEmpty()) {
-            Hashtag hashtag = HashtagMapper.instance.toEntity(hashtagDto);
-            MemberHashtag memberHashtag = MemberHashtag.builder().hashtag(hashtag).member(member).build();
-            hashTagRepository.save(hashtag);
-            return HashtagMapper.instance.toRes(memberHashtagRepository.save(memberHashtag));
+            MemberHashtag memberHashtag = HashtagMapper.toMemberHashtag(hashtagDto, member);
+            return HashtagDto.from(memberHashtagRepository.save(memberHashtag));
         } else {
             Hashtag hashtag = data.get();
             if (memberHashtagRepository.findByHashtag(hashtag).isEmpty()) {
-                MemberHashtag memberHashtag = MemberHashtag.builder().hashtag(hashtag).member(member).build();
-                return HashtagMapper.instance.toRes(memberHashtagRepository.save(memberHashtag));
+                MemberHashtag memberHashtag = HashtagMapper.toMemberHashtag(hashtag, member, hashtagDto.getColor());
+                return HashtagDto.from(memberHashtagRepository.save(memberHashtag));
             } else {
                 throw new DuplicateKeyException(ExceptionMessage.DUPLICATE.getMessage());
             }
         }
     }
 
-    public List<MemberHashtag> findByMemberIdWithType(final Long memberId, final int type) {
-        return memberHashtagRepositorySupport.findByMemberIdWithType(memberId, type);
+    public List<HashtagDto> findByMemberIdWithType(final Long memberId, final String type) {
+        List<MemberHashtag> memberHashtagList = memberHashtagRepositorySupport.findByMemberIdWithType(memberId, type);
+        return HashtagDto.from(memberHashtagList);
     }
 
     public void updateUse(final Long memberHashtagId, final Long memberId) {
         MemberHashtag memberHashtag = checkHashtag(memberHashtagId, memberId);
         memberHashtag.setUse(!memberHashtag.isUse());
-        HashtagMapper.instance.toRes(memberHashtagRepository.save(memberHashtag));
+        memberHashtagRepository.save(memberHashtag);
     }
 
     public void delete(final Long memberHashtagId, final Long memberId) {
@@ -64,6 +64,9 @@ public class HashtagService {
     }
 
     public void deleteAllByMemberId(final Long memberId) {
+        if (memberHashtagRepository.findByMemberId(memberId).isEmpty()) {
+            throw new NotFoundException(ExceptionMessage.NOT_FOUND.getMessage());
+        }
         memberHashtagRepository.deleteAllByMemberId(memberId);
     }
